fix: 修复运营中心签到配置保存问题

+++ backend/internal/handler/dto/settings.go
@@ -253,8 +253,8 @@ type SystemSettings struct {
 	DailyCheckinEnabled             bool                                   `json:"daily_checkin_enabled"`
 	DailyCheckinRequiredUsageUSD    float64                                `json:"daily_checkin_required_usage_usd"`
 	DailyCheckinUsageScope          string                                 `json:"daily_checkin_usage_scope"`
-	DailyCheckinRewardMinUSD        int                                    `json:"daily_checkin_reward_min_usd"`
-	DailyCheckinRewardMaxUSD        int                                    `json:"daily_checkin_reward_max_usd"`
+	DailyCheckinRewardMinUSD        float64                                `json:"daily_checkin_reward_min_usd"`
+	DailyCheckinRewardMaxUSD        float64                                `json:"daily_checkin_reward_max_usd"`
 	DailyCheckinDailyBudgetUSD      float64                                `json:"daily_checkin_daily_budget_usd"`
 	DailyCheckinMonthlyBudgetUSD    float64                                `json:"daily_checkin_monthly_budget_usd"`
 	DailyCheckinUserMonthlyLimitUSD float64                                `json:"daily_checkin_user_monthly_limit_usd"`
